% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{migrate}
\alias{migrate}
\title{Run migrations}
\usage{
migrate(dir, name = DBI::Id(schema = "public", table = "migrations"))
}
\arguments{
\item{dir}{Migration directory}

\item{name}{Migration table}
}
\description{
Creates the migration table if it does not already exist. Assumes that the schema already exists; fails otherwise.
}
\details{
Intersects the migration table with the migration directory. Runs the missing migration scripts in lexicographic order. Implementation relies on \link{DBI} package and its functional interface.
}
