% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{postgres.default.db.pool}
\alias{postgres.default.db.pool}
\title{Default PostgreSQL database pool}
\usage{
postgres.default.db.pool(...)
}
\arguments{
\item{...}{Extra arguments for
\code{RPostgres::\link[RPostgres]{postgresDefault}}}
}
\description{
Lazily creates a default pool for PostgreSQL connections. Call this first at
some point in your application's initialisation procedure. Raises an error if
a new default pool cannot connect to the database.
}
