% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{waitForNotify}
\alias{waitForNotify}
\title{Wait for any PostgreSQL notification}
\usage{
waitForNotify(...)
}
\arguments{
\item{...}{Extra arguments for PostgreSQL notification wait,
\code{RPostgres::\link[RPostgres]{postgresWaitForNotify}}.
Available parameters include:
\itemize{
\item timeout Seconds to wait, one by default.
}}
}
\description{
Uses the default pool. The default pool uses the default PostgreSQL
connection.
}
