% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{with.transaction}
\alias{with.transaction}
\title{Call function with pooled PostgreSQL connection}
\usage{
\method{with}{transaction}(func)
}
\arguments{
\item{func}{Function called with one database connection argument}
}
\description{
Call function with pooled PostgreSQL connection
}
