% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{withConnection}
\alias{withConnection}
\title{Call function with pooled connection}
\usage{
withConnection(what)
}
\arguments{
\item{what}{Called with one database connection argument}
}
\description{
Call function with pooled connection
}
